+++ web/packages/teleport/src/UnifiedResources/UnifiedResources.tsx
@@ -54,9 +54,9 @@ import {
   useSamlAppAction,
   SamlAppActionProvider,
 } from 'teleport/SamlApplications/useSamlAppActions';
+import { ServersideSearchPanel } from 'teleport/components/ServersideSearchPanel';
 
 import { ResourceActionButton } from './ResourceActionButton';
-import SearchPanel from './SearchPanel';
 
 export function UnifiedResources() {
   const { clusterId, isLeafCluster } = useStickyClusterId();
@@ -306,8 +306,8 @@ export function ClusterResources({
                 )}
               </Flex>
             </FeatureHeader>
-            <Flex alignItems="center" justifyContent="space-between">
-              <SearchPanel
+            <Flex alignItems="center" justifyContent="space-between" mb={3}>
+              <ServersideSearchPanel
                 params={params}
                 pathname={pathname}
                 replaceHistory={replaceHistory}
